Refactor : FE 로그인 페이지 통합 (LoginPage 삭제, /login → HomePage)
