--- com.peopleground.sagwim.fe/src/pages/LoginPage.module.css
@@ -1,83 +1,0 @@
-/* ── LoginPage ── */
-
-.root {
-  min-height: 100svh;
-  display: flex;
-  align-items: center;
-  justify-content: center;
-  padding: var(--sp-6);
-  position: relative;
-  overflow: hidden;
-}
-
-.root::before {
-  content: '';
-  position: absolute;
-  inset: 0;
-  background:
-    radial-gradient(ellipse 60% 50% at 80% 20%, color-mix(in srgb, var(--clr-accent) 8%, transparent), transparent 55%),
-    radial-gradient(ellipse 50% 50% at 20% 80%, color-mix(in srgb, #8b5cf6 6%, transparent), transparent 55%);
-  pointer-events: none;
-}
-
-.card {
-  width: 100%;
-  max-width: 420px;
-  padding: var(--sp-10) var(--sp-8);
-}
-
-.heading {
-  font-size: 1.75rem;
-  font-weight: 700;
-  margin-bottom: var(--sp-1);
-}
-
-.subheading {
-  color: var(--clr-text-muted);
-  font-size: 0.875rem;
-  margin-bottom: var(--sp-8);
-}
-
-.footer {
-  text-align: center;
-  margin-top: var(--sp-5);
-  color: var(--clr-text-muted);
-  font-size: 0.875rem;
-}
-
-.footer a {
-  color: var(--clr-accent);
-  font-weight: 600;
-}
-
-.backLink {
-  display: inline-flex;
-  align-items: center;
-  gap: var(--sp-1);
-  color: var(--clr-text-muted);
-  font-size: 0.875rem;
-  margin-bottom: var(--sp-6);
-  transition: color var(--t-fast);
-}
-
-.backLink:hover {
-  color: var(--clr-text);
-  opacity: 1;
-}
-
-.divider {
-  display: flex;
-  align-items: center;
-  gap: var(--sp-3);
-  margin: var(--sp-4) 0;
-  color: var(--clr-text-muted);
-  font-size: 0.8125rem;
-}
-
-.divider::before,
-.divider::after {
-  content: '';
-  flex: 1;
-  height: 1px;
-  background: var(--clr-border);
-}
